refactor(li): deferred-format validation hooks + Index-typed numrows/numcols + lint expansion
Three coherent threads land together because they share the same goal: kill
unconditional std::string / std::format / static_cast allocations from the
LP-build hot paths and codify the patterns so future regressions get
caught at lint time.

LP validation hot path
  - `LpValidationStats::note_bound`/`note_rhs`/`note_obj`/`note_coeff`/`note_filtered`
    gain inline template overloads that take any formattable `Loc` (ColIndex,
    RowIndex, etc.) and skip `std::format` until AFTER the threshold check
    fires — well-conditioned bounds/RHS/obj never allocate a location string.
  - All `m_validation_stats_.note_*` call sites in `linear_interface.cpp`
    drop their pre-built `std::string`/`as_label("col"|"row", idx)` fallbacks
    and pass the strong index directly.  No more "name OR fallback" branch.
  - `note_coeff(double, RowLoc, ColLoc, cfg)` 2-axis overload removes the
    `std::format("row {} col {}", ...)` allocation in `set_coeff_raw`.

Index-typed LP dimensions
  - `LinearInterface::get_numrows()` / `get_numcols()` now return `Index`
    (signed int32_t) instead of `size_t`, matching `LinearProblem`'s API
    and the solver C backends.  Closes a same-name-different-type asymmetry.
  - `m_cached_numrows_` / `m_cached_numcols_` / `m_base_numrows_` retyped
    to `Index` to match.
  - `ScaledView` ctor takes `Index` for LP-dimension parameters (n, nlo,
    nup) — single point of `int → size_t` conversion inside the std::span
    paren-init, no caller-side casts.
  - `LpStats::num_vars` / `num_constraints` retyped to `Index`.
  - 21+ redundant `static_cast<int>(strong_index)` and
    `static_cast<size_t>(get_num_*())` patterns dropped across
    `linear_interface.*`, `cascade_method.cpp`, `sddp_forward_pass.cpp`,
    `sddp_cut_store.cpp`, `sddp_cut_csv.cpp`.
  - Pass-through getters (`base_numrows`, `flatten_col_count`,
    `flatten_row_count`, `get_algorithm`, `get_threads`, `get_presolve`,
    `get_log_level`, the five `*_use_count` accessors) now return `auto`
    so the type tracks the underlying member without restating it.

Linter rewrite + 5 new rules
  - `tools/lint_strong_types.sh` → `tools/lint_strong_types.py` —
    multi-line context, comment-aware, per-rule fix hints, easier to
    extend.  Rules 1–7 preserved, behaviour unchanged.
  - Rule 8: redundant `static_cast<int>(strong_index)` and
    `static_cast<int>(*.get_numrows()/get_numcols())`.
  - Rule 9: `std::format(..., static_cast<int>(strong_index))` — strong
    indices and UidOf<Tag> are formattable directly.
  - Rule 10: `spdlog::level(std::format(...))` — defeats spdlog lazy
    formatting (allocates before the level filter has a chance to skip).
  - Rule 11: redundant `static_cast<int64_t>(*.use_count())` — `long`
    widens to `int64_t` implicitly on every supported platform.
  - Rule 12: redundant `static_cast<size_t>(container.size())` —
    `.size()` already returns `size_type = size_t`.

Compiler warnings
  - `cmake_local/CompilerWarnings.cmake` adds `-Wno-maybe-uninitialized`
    to the GCC-specific list.  At `-O2`/`-O3` the warning hallucinates
    "may be used uninitialized" through libstdc++'s `std::string` /
    `std::optional` aggregate-init paths, breaking Release builds under
    `-Werror`.  Clang's flow analysis doesn't have this problem.

All 3097 unit tests pass.  Linter clean on the new rules.

+++ source/cascade_method.cpp
@@ -276,7 +276,7 @@ void CascadePlanningMethod::add_elastic_targets(
           .context = p.context,
       });
     }
-    const auto first_slack_col = static_cast<int>(li.get_numcols());
+    const auto first_slack_col = li.get_numcols();
     li.add_cols(slack_cols);
 
     // Mirror the slacks into the persistent `m_dynamic_cols_` registry
@@ -348,7 +348,7 @@ void CascadePlanningMethod::clear_all_cuts(PlanningLP& planning_lp,
     {
       auto& li = planning_lp.system(scene, phase).linear_interface();
       const auto base = static_cast<int>(states[phase].base_nrows);
-      const auto current = static_cast<int>(li.get_numrows());
+      const auto current = li.get_numrows();
 
       if (current > base) {
         auto indices = std::ranges::to<std::vector>(iota_range(base, current));
